+++ src/index.php
@@ -1,5 +1,12 @@
 <?php
+/**
+ * Test Composer plugin for testing monorepo workflows
+ *
+ * @package itcig/composer-plugin-sandbox
+ */
 
 namespace Cig\Test;
 
-define('PACKAGE_VERSION', '0.1.0');
+define('PACKAGE_VERSION', '0.3.1');
+
+echo 'new test';