+++ CHANGELOG.md
@@ -1,7 +1,172 @@
 # Change Log
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [0.3.1](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.3.1-alpha.0...@itcig/composer-plugin-sandbox@0.3.1) (2021-10-25)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [0.3.1-alpha.0](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.3.0...@itcig/composer-plugin-sandbox@0.3.1-alpha.0) (2021-10-21)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+# [0.3.0](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.3.0-alpha.1...@itcig/composer-plugin-sandbox@0.3.0) (2021-10-21)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+# [0.3.0-alpha.1](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.3.0-alpha.0...@itcig/composer-plugin-sandbox@0.3.0-alpha.1) (2021-10-21)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+# [0.3.0-alpha.0](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.8...@itcig/composer-plugin-sandbox@0.3.0-alpha.0) (2021-10-21)
+
+### Features
+
+- test minor update on composer plugin
+  ([87163f0](https://github.com/itcig/composer-plugin-sandbox/commit/87163f0bc0c841daabd4ce2e0f7f8d6c18870650))
+
+## [0.2.8](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.8-alpha.12...@itcig/composer-plugin-sandbox@0.2.8) (2021-10-20)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.8-alpha.12](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.8-alpha.11...@itcig/composer-plugin-sandbox@0.2.8-alpha.12) (2021-10-20)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.8-alpha.11](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.8-alpha.10...@itcig/composer-plugin-sandbox@0.2.8-alpha.11) (2021-10-20)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.8-alpha.10](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.8-alpha.9...@itcig/composer-plugin-sandbox@0.2.8-alpha.10) (2021-10-20)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.8-alpha.9](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.8-alpha.8...@itcig/composer-plugin-sandbox@0.2.8-alpha.9) (2021-10-20)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.8-alpha.8](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.8-alpha.7...@itcig/composer-plugin-sandbox@0.2.8-alpha.8) (2021-10-20)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.8-alpha.7](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.8-alpha.6...@itcig/composer-plugin-sandbox@0.2.8-alpha.7) (2021-10-20)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.8-alpha.6](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.8-alpha.5...@itcig/composer-plugin-sandbox@0.2.8-alpha.6) (2021-10-20)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.8-alpha.5](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.8-alpha.4...@itcig/composer-plugin-sandbox@0.2.8-alpha.5) (2021-10-20)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.8-alpha.4](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.8-alpha.3...@itcig/composer-plugin-sandbox@0.2.8-alpha.4) (2021-10-20)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.8-alpha.3](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.8-alpha.2...@itcig/composer-plugin-sandbox@0.2.8-alpha.3) (2021-10-20)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.8-alpha.2](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.8-alpha.1...@itcig/composer-plugin-sandbox@0.2.8-alpha.2) (2021-10-20)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.8-alpha.1](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.8-alpha.0...@itcig/composer-plugin-sandbox@0.2.8-alpha.1) (2021-10-19)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.8-alpha.0](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.7...@itcig/composer-plugin-sandbox@0.2.8-alpha.0) (2021-10-19)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.7](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.7-alpha.0...@itcig/composer-plugin-sandbox@0.2.7) (2021-10-19)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.7-alpha.0](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.0...@itcig/composer-plugin-sandbox@0.2.7-alpha.0) (2021-10-19)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [0.2.7-dev.3](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.7-dev.2...@itcig/composer-plugin-sandbox@0.2.7-dev.3) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.7-dev.2](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.7-dev.1...@itcig/composer-plugin-sandbox@0.2.7-dev.2) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.7-dev.1](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.7-dev.0...@itcig/composer-plugin-sandbox@0.2.7-dev.1) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.7-dev.0](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.6...@itcig/composer-plugin-sandbox@0.2.7-dev.0) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.6](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.6-dev.7...@itcig/composer-plugin-sandbox@0.2.6) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.6-dev.7](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.6-dev.6...@itcig/composer-plugin-sandbox@0.2.6-dev.7) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.6-dev.6](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.6-dev.5...@itcig/composer-plugin-sandbox@0.2.6-dev.6) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.6-dev.5](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.6-dev.4...@itcig/composer-plugin-sandbox@0.2.6-dev.5) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.6-dev.4](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.6-dev.3...@itcig/composer-plugin-sandbox@0.2.6-dev.4) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.6-dev.3](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.6-dev.2...@itcig/composer-plugin-sandbox@0.2.6-dev.3) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.6-dev.2](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.6-dev.1...@itcig/composer-plugin-sandbox@0.2.6-dev.2) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.6-dev.1](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.6-dev.0...@itcig/composer-plugin-sandbox@0.2.6-dev.1) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.6-dev.0](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.5...@itcig/composer-plugin-sandbox@0.2.6-dev.0) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.5](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.0...@itcig/composer-plugin-sandbox@0.2.5) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.4](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.0...@itcig/composer-plugin-sandbox@0.2.4) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.3](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.0...@itcig/composer-plugin-sandbox@0.2.3) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
+
+## [0.2.2](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.0...@itcig/composer-plugin-sandbox@0.2.2) (2021-10-15)
+
+**Note:** Version bump only for package @itcig/composer-plugin-sandbox
 
 ## [0.2.1](https://github.com/itcig/composer-plugin-sandbox/compare/@itcig/composer-plugin-sandbox@0.2.0...@itcig/composer-plugin-sandbox@0.2.1) (2021-10-14)